# docs/swagger.yaml
definitions:
...
  schema.TokenResponse:
    properties:
      accessToken:
        type: string
      expiresIn:
        type: integer
      refreshToken:
        type: string
      tokenType:
        type: string
    type: object
...
  shared.HTTPError:
    properties:
      code:
        type: string
      details:
        type: object
      message:
        type: string
    type: object
info:
  contact: {}
paths:
...
  /auth/login:
    post:
      description: Refresh access token
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/schema.TokenResponse'
        "403":
          description: Forbidden
          schema:
            $ref: '#/definitions/shared.HTTPError'
      summary: Refresh access token
      tags:
      - auth
...
swagger: "2.0"